Conflicts:
	producers/miniTLU/include/tlu/miniTLUController.hh
	producers/miniTLU/src/miniTLUControl.cxx
	producers/miniTLU/src/miniTLU_Producer.cxx

solved conflicts. Most were due to use of shared_ptr vs unique_ptr

+++ main/include/eudaq/DetectorEvent.hh
@@ -14,7 +14,7 @@ namespace eudaq {
     EUDAQ_DECLARE_EVENT(DetectorEvent);
     public:
     virtual void Serialize(Serializer &) const;
-    explicit DetectorEvent(unsigned runnumber, unsigned eventnumber, unsigned long long timestamp) :
+    explicit DetectorEvent(unsigned runnumber, unsigned eventnumber, uint64_t timestamp) :
       Event(runnumber, eventnumber, timestamp)
     {}
     //     explicit DetectorEvent(const TLUEvent & tluev) :
